feat: useGPTとしてhook化.
